llm-evaluators/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
appnope==0.1.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
backoff==2.2.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
comm==0.2.2
debugpy==1.8.7
decorator==5.1.1
distro==1.9.0
docker-pycreds==0.4.0
docstring_parser==0.16
emoji==2.14.0
exceptiongroup==1.2.2
executing==2.1.0
frozenlist==1.5.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
gql==3.5.0
graphql-core==3.2.5
grpcio==1.67.0
grpcio-status==1.67.0
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
idna==3.10
instructor==1.6.3
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
jupyter_client==8.6.3
jupyter_core==5.7.2
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.1.0
nest-asyncio==1.6.0
numpy==2.1.2
openai==1.52.2
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.3
rsa==4.9
sentry-sdk==2.17.0
set-env-colab-kaggle-dotenv==0.1.3
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
stack-data==0.6.3
tenacity==9.0.0
tornado==6.4.1
tqdm==4.66.6
traitlets==5.14.3
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uuid_utils==0.9.0
uv==0.4.27
wandb==0.18.5
wcwidth==0.2.13
weave==0.51.17
yarl==1.16.0
